requirements.txt
# Python 3.11
requests>=2.28.0
beautifulsoup4>=4.11.0
playwright>=1.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
